% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondsFundamentalsAPI.R
\name{get_bonds_fundamentals_data}
\alias{get_bonds_fundamentals_data}
\title{Get bonds fundamentals data}
\usage{
get_bonds_fundamentals_data(api_token, isin)
}
\arguments{
\item{api_token}{The API token for authentication.}

\item{isin}{The ISIN (International Securities Identification Number) of the bond.}
}
\value{
A list containing the bond fundamentals data.
}
\description{
This function retrieves bond fundamentals data from the API.
}
\examples{
api_token <- "demo"
isin <- "DE000CB83CF0"
result <- get_bonds_fundamentals_data(api_token, isin)

}
