% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradingHoursStockMarketHolidaysAPI.R
\name{get_details_trading_hours_stock_market_holidays}
\alias{get_details_trading_hours_stock_market_holidays}
\title{Get details trading hours, stock market holidays and symbol change history}
\usage{
get_details_trading_hours_stock_market_holidays(
  api_token,
  code,
  from_date = NULL,
  to_date = NULL
)
}
\arguments{
\item{api_token}{The API token for authentication.}

\item{code}{Use the exchange code from the API endpoint}

\item{from_date}{the format is ‘YYYY-MM-DD’}

\item{to_date}{the format is ‘YYYY-MM-DD’}
}
\value{
A list containing the details trading hours, stock market holidays and symbol change history.
}
\description{
This function retrieves details trading hours, stock market holidays and symbol change history from the API.
}
\examples{
api_token <- "demo"
code <- "US"
from_date <- "2017-09-10"
to_date <- "2017-09-12"

result <- get_details_trading_hours_stock_market_holidays(api_token, code, from_date, to_date)

}
