% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarningTrendsAPI.R
\name{get_earning_trends_data}
\alias{get_earning_trends_data}
\title{Get earning trends data}
\usage{
get_earning_trends_data(api_token, symbols)
}
\arguments{
\item{api_token}{The API token for authentication.}

\item{symbols}{You can request specific symbols to get historical and upcoming data}
}
\value{
A list containing earning trends.
}
\description{
This function retrieves earning trends data from the API.
}
\examples{
api_token <- "demo"
symbols <- "AAPL.US"
result <- get_earning_trends_data(api_token, symbols)

}
