% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FundamentalDataAPI.R
\name{get_fundamentals_data}
\alias{get_fundamentals_data}
\title{Get fundamentals data data}
\usage{
get_fundamentals_data(api_token, ticker)
}
\arguments{
\item{api_token}{The API token for authentication.}

\item{ticker}{consists of two parts: [SYMBOL_NAME].[EXCHANGE_ID]}
}
\value{
A list containing the fundamental data.
}
\description{
This function retrieves fundamentals data from the API.
}
\examples{
api_token <- "demo"
ticker <- "AAPL.US"
result <- get_fundamentals_data(api_token, ticker)

}
