% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoricalSplitsAPI.R
\name{get_historical_splits_data}
\alias{get_historical_splits_data}
\title{Get historical splits data}
\usage{
get_historical_splits_data(api_token, ticker, date_from = NULL, date_to = NULL)
}
\arguments{
\item{api_token}{The API token for authentication.}

\item{ticker}{consists of two parts: [SYMBOL_NAME].[EXCHANGE_ID].}

\item{date_from}{date from with format Y-m-d}

\item{date_to}{date from with format Y-m-d}
}
\value{
A list containing historical splits data.
}
\description{
This function retrieves historical splits data from the API.
}
\examples{
api_token <- "demo"
ticker <- "AAPL.US"
date_from <- "2017-09-10"
date_to <- "2017-09-12"
result <- get_historical_splits_data(api_token, ticker, date_from, date_to)

}
