% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListOfExchangerAPI.R
\name{get_list_of_exchanges}
\alias{get_list_of_exchanges}
\title{Get list of exchanges}
\usage{
get_list_of_exchanges(api_token)
}
\arguments{
\item{api_token}{The API token for authentication.}
}
\value{
A list containing the list of exchanges.
}
\description{
This function retrieves list of exchanges from the API.
}
\examples{
api_token <- "demo"
result <- get_list_of_exchanges(api_token)

}
