% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiveStockPricesAPI.R
\name{get_live_stock_prices}
\alias{get_live_stock_prices}
\title{Get live stock prices}
\usage{
get_live_stock_prices(api_token, ticker, s = NULL)
}
\arguments{
\item{api_token}{The API token for authentication.}

\item{ticker}{consists of two parts: [SYMBOL_NAME].[EXCHANGE_ID]}

\item{s}{add “s=” parameter to your function and you will be able to get data for multiple tickers at one request, all tickers should be separated with a comma.}
}
\value{
A list containing a live stock prices.
}
\description{
This function retrieves live stock prices from the API.
}
\examples{
api_token <- "demo"
ticker <- "AAPL.US"
s <- "VTI,EUR.FOREX"
result <- get_live_stock_prices(api_token, ticker, s)

}
