% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentimentAPI.R
\name{get_sentiment_data}
\alias{get_sentiment_data}
\title{Get sentiment data.}
\usage{
get_sentiment_data(api_token, s, from_date = NULL, to_date = NULL)
}
\arguments{
\item{api_token}{The API token for authentication.}

\item{s}{Parameter to your URL and you will be able to get data for multiple tickers at one request, all tickers should be separated with a comma.}

\item{from_date}{Format: YYYY-MM-DD.}

\item{to_date}{Format: YYYY-MM-DD.}
}
\value{
A list containing the sentiment data.
}
\description{
This function retrieves sentiment data from the API.
}
\examples{
api_token <- "demo"
from_date <- "2017-09-10"
to_date <- "2017-09-12"
s <- "AAPL.MX"
result <- get_sentiment_data(api_token, s, from_date, to_date)

}
