% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymbolsChangeHistoryAPI.R
\name{symbol_change_history}
\alias{symbol_change_history}
\title{Get symbol change history}
\usage{
symbol_change_history(api_token, from_date = NULL, to_date = NULL)
}
\arguments{
\item{api_token}{The API token for authentication.}

\item{from_date}{The start date for symbol change history (optional).}

\item{to_date}{The end date for symbol change history (optional).}
}
\value{
A list containing symbol change history data.
}
\description{
This function retrieves symbol change history from the API.
}
\examples{
api_token <- "demo"
from_date <- "2023-01-01"
to_date <- "2023-12-31"
result <- symbol_change_history(api_token, from_date, to_date)

}
